% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Packager.R
\name{Packager}
\alias{Packager}
\title{An R6 class for assisting with the packaging of "Io" x "DataDAG" formatted data packages}
\description{
An R6 class for assisting with the packaging of "Io" x "DataDAG" formatted data packages
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{recipe}}{Input Io recipe.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Packager$new()}}
\item \href{#method-build_package}{\code{Packager$build_package()}}
\item \href{#method-clone}{\code{Packager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new "Packager" instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Packager$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{Packager} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-build_package"></a>}}
\if{latex}{\out{\hypertarget{method-build_package}{}}}
\subsection{Method \code{build_package()}}{
Builds Io/DataDAG data package.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Packager$build_package(pkg_dir, data, row_metadata = NULL, col_metadata = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pkg_dir}}{Path where data package should be generated}

\item{\code{data}}{data.frame Main dataset}

\item{\code{row_metadata}}{data.frame (Optional) row metadata}

\item{\code{col_metadata}}{data.frame (Optional) col metadata}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Packager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
