% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Packager.R
\name{Packager}
\alias{Packager}
\title{An R6 class for assisting with creating and interacting with "eco" data packages
packages}
\description{
An R6 class for assisting with creating and interacting with "eco" data packages
packages
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{recipe}}{Input eco recipe.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Packager$new()}}
\item \href{#method-build_package}{\code{Packager$build_package()}}
\item \href{#method-update_package}{\code{Packager$update_package()}}
\item \href{#method-clone}{\code{Packager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new "Packager" instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Packager$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{Packager} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-build_package"></a>}}
\if{latex}{\out{\hypertarget{method-build_package}{}}}
\subsection{Method \code{build_package()}}{
Builds an eco data package.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Packager$build_package(
  resources,
  annotations = c(),
  views = c(),
  node_metadata = list(),
  dag_metadata = list(),
  profile = "",
  pkg_dir = "./",
  include_summary = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resources}}{List of resources (dataframes/tibbles) to include in package}

\item{\code{annotations}}{Vector of filepaths to annotation files, or string annotations}

\item{\code{views}}{List of filepaths to vega-lite views, or list representations of such views}

\item{\code{node_metadata}}{\link{Optional} Additional node-level metadata to include,
either as a list, or path to a yml/json file; Node level metadata is \emph{not}
carried over to child nodes.}

\item{\code{dag_metadata}}{\link{Optional} Additional dag-level metadata to include,
either as a list, or path to a yml/json file; DAG-level metadata \emph{is} copied over
to child packages by default; if the same field is specified downstream, the
value of the field is updated.}

\item{\code{profile}}{\link{Optional} Name of metadata profile to use for validation, or "", for none}

\item{\code{pkg_dir}}{\link{Optional} Location where data package should be saved (default: "./")}

\item{\code{include_summary}}{\link{Optional} Whether or not to compute & embed summary statistics at the time of package creation (default: FALSE)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_package"></a>}}
\if{latex}{\out{\hypertarget{method-update_package}{}}}
\subsection{Method \code{update_package()}}{
Updates an existing eco datapackage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Packager$update_package(
  existing,
  resources,
  annotations = c(),
  views = c(),
  node_metadata = list(),
  dag_metadata = list(),
  profile = "",
  pkg_dir = "./",
  include_summary = False
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{existing}}{Path to existing datapackage.json to be extended}

\item{\code{resources}}{List of names resources to include in the data package}

\item{\code{annotations}}{Vector of filepaths to annotation files, or string annotations}

\item{\code{views}}{Vector of filepaths to vega-lite views, or dict representations of such views}

\item{\code{node_metadata}}{\link{Optional} Additional node-level metadata to include,
either as a list, or path to a yml/json file; Node level metadata is \emph{not}
carried over to child nodes.}

\item{\code{dag_metadata}}{\link{Optional} Additional dag-level metadata to include,
either as a list, or path to a yml/json file; DAG-level metadata \emph{is} copied over
to child packages by default; if the same field is specified downstream, the
value of the field is updated.}

\item{\code{profile}}{\link{Optional} Name of metadata profile to use for validation, or "", for none}

\item{\code{pkg_dir}}{\link{Optional} Location where data package should be saved (default: "./")}

\item{\code{include_summary}}{\link{Optional} Whether or not to compute & embed summary statistics at the time of package creation (default: False)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Packager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
